+++ eslint.config.js
@@ -1,14 +1,14 @@
-const js = require('@eslint/js')
+import js from '@eslint/js'
 
-module.exports = [
+export default [
   {
     ignores: ['node_modules/'],
   },
   {
     files: ['**/*.js'],
     languageOptions: {
-      ecmaVersion: 2024,
-      sourceType: 'commonjs',
+      ecmaVersion: 2020,
+      sourceType: 'module',
       globals: {
         console: 'readonly',
         process: 'readonly',
@@ -16,9 +16,6 @@ module.exports = [
         __dirname: 'readonly',
         __filename: 'readonly',
         global: 'readonly',
-        require: 'readonly',
-        exports: 'writable',
-        module: 'readonly',
         setImmediate: 'readonly',
         clearImmediate: 'readonly',
         setInterval: 'readonly',
@@ -29,19 +26,15 @@ module.exports = [
     },
     rules: {
       ...js.configs.recommended.rules,
-      quotes: ['error', 'single', { avoidEscape: true, allowTemplateLiterals: true }],
-      semi: ['error', 'never'],
+      'quotes': ['error', 'single', { avoidEscape: true, allowTemplateLiterals: true }],
+      'semi': ['error', 'never'],
       'semi-style': ['error', 'last'],
       'no-unused-vars': ['error', { args: 'none' }],
-      'array-bracket-spacing': [
-        'warn',
-        'always',
-        { arraysInArrays: false, objectsInArrays: false },
-      ],
+      'array-bracket-spacing': ['warn', 'always', { arraysInArrays: false, objectsInArrays: false }],
       'no-trailing-spaces': ['error'],
       'space-before-function-paren': ['error', 'always'],
       'object-curly-spacing': ['error', 'always'],
-      indent: ['error', 2, { SwitchCase: 1 }],
+      'indent': ['error', 2, { SwitchCase: 1 }],
       'key-spacing': [
         'warn',
         {